FoodFinder 1st commit
Starting with the Jax-RS-Heroku template, this application seeks to host
a means for a client to make API calls and receive back a list of food
trucks near a given location.  While FoodFinder.java is equipped to
process GET requests as a RESTful API, the current implementation simply
uses FoodFrame to show functionality in a runnable Java applet.